// src/dorc-web/src/apis/dorc-api/models/PermissionDto.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface PermissionDto
 */
export interface PermissionDto {
  /**
   * @type {number}
   * @memberof PermissionDto
   */
  Id?: number;
  /**
   * @type {string}
   * @memberof PermissionDto
   */
  PermissionName?: string | null;
  /**
   * @type {string}
   * @memberof PermissionDto
   */
  DisplayName?: string | null;
}