// src/dorc-web/src/apis/dorc-api/models/UserPermDto.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface UserPermDto
 */
export interface UserPermDto {
  /**
   * @type {number}
   * @memberof UserPermDto
   */
  Id?: number;
  /**
   * @type {string}
   * @memberof UserPermDto
   */
  Database?: string | null;
  /**
   * @type {string}
   * @memberof UserPermDto
   */
  Role?: string | null;
  /**
   * @type {string}
   * @memberof UserPermDto
   */
  User?: string | null;
}